+++ Cargo.toml
@@ -25,29 +25,44 @@
 [dependencies]
     anyhow="1.0"
     glam="0.29.0"
-    glium={version="0.33", optional=true}
+    log="0.4"
+    rayon="1.10.0"
+    winit="0.30.0"
+
+[target.'cfg(target_arch = "x86_64")'.dependencies]
+    env_logger="0.11"
+    glium="0.33"
     petgraph="0.6.5"
     pollster="0.3"
     rand="0.8.5"
-    rayon="1.10.0"
-    winit={version="0.28.7", optional=true}
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+    console_error_panic_hook="0.1.6"
+    console_log="1.0"
+    wasm-bindgen="0.2.100"
+    wasm-bindgen-futures="0.4.50"
+
+    [dependencies.web-sys]
+        version="0.3.77"
+        # https://docs.rs/web-sys/latest/web_sys/#modules
+        features=["Document", "Element", "Window"]
 
     [dependencies.wgpu]
         default-features=false
-        features=["webgpu", "static-dxc", "web", "std", "parking_lot"]
-        optional=true
+        features=[
+            "webgl",
+            "static-dxc",
+            "web",
+            "std",
+            "wgsl",
+            "parking_lot",
+            # webgpu # when API is stable
+        ]
         version="26.0.1"
-
-    [dependencies.smallvec]
-        default-features=false
-        features=["union"]
-        optional=true
-        version="1.15.1"
 
     [dependencies.specs]
         default-features=false
         features=["specs-derive", "parallel", "stdweb"]
-        optional=true
         version="0.20.0"
 
 
@@ -56,10 +71,6 @@
     petgraph-gen="0.1.3"
     serde={version="1.0", features=["derive"]}
     serde_json="1.0"
-
-[features]
-    desktop=["dep:glium", "dep:winit"]
-    wasm=["dep:smallvec", "dep:specs", "dep:wgpu"]
 
 [[example]]
     name="wikipedia"
