+++ examples/ecs.rs
@@ -1,5 +1,5 @@
 use glam::Vec2;
-use grapher::wasm::simulator::Simulator;
+use grapher::web::simulator::Simulator;
 
 fn build_graph() -> (Vec<Vec2>, Vec<Vec2>) {
     let nodes = vec![Vec2::new(0.0, 0.0), Vec2::new(0.0, 1.0)];
