+++ src/native/properties.rs
