--- src/wasm.rs
@@ -1,3 +1,0 @@
-mod quadtree_ecs;
-pub mod renderer_ecs;
-pub mod simulator;
