+++ src/web.rs
@@ -1,0 +1,5 @@
+mod quadtree;
+mod renderer;
+pub mod simulator;
+
+pub use renderer::run_web;
